LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/overflow-clip-scroll-size-expected.txt


PASS scroll size should match that of size specified by overflow: clip
FAIL scroll size should take into account border size and overflow-clip-margin assert_equals: expected 105 but got 104
